+++ Plugin/src/Plugin.cs
@@ -1,16 +1,34 @@
-﻿using System.Reflection;
+﻿using System;
+using System.Collections.Generic;
+using System.Reflection;
 using UnityEngine;
 using BepInEx;
-using LethalLib.Modules;
 using System.IO;
+using BepInEx.Configuration;
+using LethalLib.Modules;
 
-namespace Skull
+namespace SkullEnemy
 {
     [BepInPlugin(PluginInfo.PLUGIN_GUID, PluginInfo.PLUGIN_NAME, PluginInfo.PLUGIN_VERSION)]
-    public class SkullPlugin : BaseUnityPlugin
+    public class SkullEnemyPlugin : BaseUnityPlugin
     {
+        public static SkullEnemyPlugin Instance;
+
+        public ConfigEntry<float> ConfigMovementSpeed;
+        public ConfigEntry<float> ConfigRotationSpeed;
+        public ConfigEntry<bool> ConfigCanOnlyKillTargetPlayer;
+
+        private static readonly Dictionary<Levels.LevelTypes, int> DefaultLevelRarities = new()
+        {
+            [Levels.LevelTypes.RendLevel] = 66,
+            [Levels.LevelTypes.TitanLevel] = 66,
+            [Levels.LevelTypes.DineLevel] = 66
+        };
+
         private void Awake()
         {
+            Instance = this;
+
             var assemblyLocation = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
             if (assemblyLocation.IsNullOrWhiteSpace())
             {
@@ -24,17 +42,31 @@ namespace Skull
                 return;
             }
 
-            var skullEnemy = assetBundle.LoadAsset<EnemyType>("SkullEnemy");
-            var terminalNode = assetBundle.LoadAsset<TerminalNode>("SkullTerminalNode");
-            var terminalKeyword = assetBundle.LoadAsset<TerminalKeyword>("SkullTerminalKeyword");
+            ConfigMovementSpeed = Config.Bind("General", "MovementSpeed", 2f,
+                "The speed at which the enemy moves towards its target.");
+            ConfigRotationSpeed = Config.Bind("General", "RotationSpeed", 2f,
+                "The speed at which the enemy rotates towards its target.");
+            ConfigCanOnlyKillTargetPlayer = Config.Bind("General", "CanOnlyKillTargetPlayer", true,
+                "If enabled, the enemy will only be able to kill its target player.");
 
-            NetworkPrefabs.RegisterNetworkPrefab(skullEnemy.enemyPrefab);
+            var levelRarities = new Dictionary<Levels.LevelTypes, int>();
+            foreach (Levels.LevelTypes levelType in Enum.GetValues(typeof(Levels.LevelTypes)))
+            {
+                if (!Levels.LevelTypes.Vanilla.HasFlag(levelType) || levelType == Levels.LevelTypes.Vanilla)
+                    continue;
 
-            Enemies.RegisterEnemy(skullEnemy, 66,
-                Levels.LevelTypes.TitanLevel | Levels.LevelTypes.DineLevel | Levels.LevelTypes.RendLevel,
-                Enemies.SpawnType.Default, terminalNode, terminalKeyword);
+                var configEntry = Config.Bind("Spawning", $"{levelType}Rarity",
+                    DefaultLevelRarities.GetValueOrDefault(levelType, 0), $"The spawn rarity on {levelType}.");
+                levelRarities.Add(levelType, configEntry.Value);
+            }
 
-            // Required by Unity Netcode Patcher
+            var skullEnemyType = assetBundle.LoadAsset<EnemyType>("SkullEnemy");
+            var skullTerminalNode = assetBundle.LoadAsset<TerminalNode>("SkullTerminalNode");
+            var skullTerminalKeyword = assetBundle.LoadAsset<TerminalKeyword>("SkullTerminalKeyword");
+
+            NetworkPrefabs.RegisterNetworkPrefab(skullEnemyType.enemyPrefab);
+            Enemies.RegisterEnemy(skullEnemyType, levelRarities, null, skullTerminalNode, skullTerminalKeyword);
+
             var types = Assembly.GetExecutingAssembly().GetTypes();
             foreach (var type in types)
             {